added user-edit selection to users/settings.html
